Init v2(Handle all channel & Remove auto comment)

+++ app/src/main/java/com/redbeanlatte11/factchecker/data/source/local/FactCheckerDataBaseMigration.kt
@@ -22,7 +22,7 @@ class FactCheckerDataBaseMigration {
                         thumbnails TEXT NOT NULL, 
                         localized TEXT NOT NULL,
                         country TEXT DEFAULT 'KR',
-                        viewCount BIGINT NOT NULL,
+                        viewCount INTEGER NOT NULL,
                         commentCount INTEGER NOT NULL,
                         subscriberCount INTEGER NOT NULL,
                         hiddenSubscriberCount INTEGER NOT NULL,
